Create most basic facial recognition functionality
This lives in the file recognise.py. It uses the face_recognition module to do most of the heavy lifting and is designed to anticipate a video stream and perform face recognition on it. Also added skeletons for a few tests to be done on the recognise file.

Also added a config file and config functionality in general as it is required to keep track of the folder where everyone's face is stored