Dev recognise faces (#7)
* Populate test db at install time and include configuration file

1. Populate test db at install by creating the file nipo/tests/populate.py and having data there that will populate the test db. This is called from setup.py(install time)

2. Create a configuration file at install which will be at the root directory and will store the project's absolute path and will store the path to the faces folder.

* Fix missing semicolon in nipo/__init__ causing tests to fail

* Create most basic facial recognition functionality

This lives in the file recognise.py. It uses the face_recognition module to do most of the heavy lifting and is designed to anticipate a video stream and perform face recognition on it. Also added skeletons for a few tests to be done on the recognise file.

Also added a config file and config functionality in general as it is required to keep track of the folder where everyone's face is stored

* Bumped sqlalchemy to v1.3.0 to deal with security vulnerability