Include initial functionality to introduce the concept of users.
Hint: Making this commit after not having seen the codebase for a few months so I dont quite remember what is going on. This commit may end up being dodgy.
Users need to exist in a db table, have usernames and passwords to be able to log in and out.