[BUGFIX] Fix issue occuring when performing multiple calls requiring DB access (#73)
When accessing the DB, there was an issue that would cause an EOF exception to be emitted
by psycopg2 if the db was accessed multiple times in the same celery instance. This is caused
by the fact that SQLAlchemy sessions are designed to be used within a single process.
This fixes this by having each process keep its own instance of the SQLAlchemy session.